more fixes and typecast removal
