+++ app/services/scene-collections/nodes/scene-items.ts
@@ -119,7 +119,7 @@ export class SceneItemsNode extends Node<ISchema, {}> {
 
     this.data.items.forEach(item => {
       if (item.sceneNodeType === 'folder') return;
-      const hotkeys = (item as ISceneItemInfo).hotkeys;
+      const hotkeys = item.hotkeys;
       if (hotkeys) promises.push(hotkeys.load({ sceneItemId: item.id }));
     });
