+++ app/services/scenes/index.ts
@@ -2,5 +2,4 @@ export * from './scenes';
 export * from './scene';
 export * from './scene-item';
 export * from './scene-folder';
-export * from './scenes-api';
 export * from './scene-node';
